mcp_yolo_server.py
#!/usr/bin/env python3
"""
MCP YOLO Server - Object detection and counting
"""

import json
import base64
import io
from typing import Dict, List, Any
from PIL import Image

# MCP server imports
from mcp.server import Server
from mcp.server.stdio import stdio_server
import mcp.types as types

# You would import your actual YOLO model here
# from ultralytics import YOLO

server = Server("yolo-server")

def decode_base64_image(b64_string: str) -> Image.Image:
    """Decode base64 string to PIL Image"""
    img_data = base64.b64decode(b64_string)
    return Image.open(io.BytesIO(img_data))

@server.tool()
async def count_objects(frames: List[str], class_name: str = "person") -> Dict[str, Any]:
    """
    Count objects in frames using YOLO
    
    Args:
        frames: List of base64-encoded PNG images
        class_name: Object class to count (person|bicycle|car|etc)
    
    Returns:
        Dict with count and bounding boxes
    """
    # Decode frames
    images = [decode_base64_image(frame) for frame in frames]
    
    # TODO: Replace with actual YOLO inference
    # model = YOLO('yolov8n.pt')
    # total_count = 0
    # all_boxes = []
    # 
    # for img in images:
    #     results = model(img)
    #     for r in results:
    #         boxes = r.boxes
    #         for box in boxes:
    #             if model.names[int(box.cls)] == class_name:
    #                 total_count += 1
    #                 all_boxes.append(box.xyxy.tolist())
    
    # Stub response
    if class_name == "person":
        count = 2
    elif class_name == "bicycle":
        count = 1
    else:
        count = 0
    
    return {
        "count": count,
        "class": class_name,
        "frames_analyzed": len(frames),
        "confidence": 0.95
    }

@server.tool()
async def detect_objects(frames: List[str], confidence_threshold: float = 0.5) -> Dict[str, Any]:
    """
    Detect all objects in frames
    
    Args:
        frames: List of base64-encoded PNG images
        confidence_threshold: Minimum confidence for detection
    
    Returns:
        Dict with all detected objects and their counts
    """
    images = [decode_base64_image(frame) for frame in frames]
    
    # TODO: Replace with actual detection
    # Stub response
    detections = {
        "person": 2,
        "bicycle": 1,
        "car": 0
    }
    
    return {
        "detections": detections,
        "total_objects": sum(detections.values()),
        "frames_analyzed": len(frames),
        "confidence_threshold": confidence_threshold
    }

async def main():
    """Run the MCP server"""
    async with stdio_server() as (read_stream, write_stream):
        await server.run(read_stream, write_stream)

if __name__ == "__main__":
    import asyncio
    asyncio.run(main())